Update add dream from a fragment to pop up alert
