--- app/src/main/java/com/example/projectlayout/ui/dream/fragment_addDream.java
@@ -1,198 +1,0 @@
-package com.example.projectlayout.ui.dream;
-
-import android.Manifest;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.drawable.BitmapDrawable;
-import android.os.Build;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ImageView;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
-import androidx.core.content.ContextCompat;
-import androidx.fragment.app.Fragment;
-import androidx.navigation.fragment.NavHostFragment;
-
-import com.example.projectlayout.R;
-
-import java.io.ByteArrayOutputStream;
-import java.util.ArrayList;
-
-import static android.app.Activity.RESULT_OK;
-
-
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link fragment_addDream#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class fragment_addDream extends Fragment {
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-
-    private static final int IMAGE_PICK_CODE = 1000;
-    private static final int PERMISSION_CODE = 1001;
-
-    ImageView mImageView;
-    private boolean recInserted;
-    private DreamDatabase db;
-    private EditText description;
-    ArrayList<Dreamboard> list;
-
-    public fragment_addDream() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment fragment_addDream.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static fragment_addDream newInstance(String param1, String param2) {
-        fragment_addDream fragment = new fragment_addDream();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-
-
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View root = inflater.inflate(R.layout.fragment_add_dream, container, false);
-
-
-
-        description = (EditText) root.findViewById(R.id.description);
-
-        db = new DreamDatabase(getActivity());
-        db.openReadable();
-
-        //imageview
-        mImageView = root.findViewById(R.id.image);
-        mImageView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                    if (ContextCompat.checkSelfPermission(getContext(), Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_DENIED) {
-                        //permission not granted, request it.
-                        String[] permissions = {Manifest.permission.READ_EXTERNAL_STORAGE};
-                        //show popup for runtime permission
-                        requestPermissions(permissions, PERMISSION_CODE);
-                    } else {
-                        //permission already granted
-                        pickImageFromGallery();
-                    }
-                } else {
-                    //system os is less then marshmallow
-                    pickImageFromGallery();
-                }
-            }
-        });
-
-        //add Dreamboard
-        Button add = (Button) root.findViewById(R.id.add);
-        add.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                if (description.getText() != null && (int) mImageView.getTag() == 1)
-                    recInserted=true;
-                    recInserted = db.addRow(description.getText().toString(), imageViewToByte(mImageView));
-
-                if (recInserted) {
-                    Toast.makeText(getActivity(), "Success add row", Toast.LENGTH_SHORT).show();
-                    NavHostFragment.findNavController(fragment_addDream.this)
-                            .navigate(R.id.action_fragment_addDream_to_nav_dream);
-                } else {
-                    Toast.makeText(getActivity(), "Not Success add row", Toast.LENGTH_SHORT).show();
-                }
-
-
-
-            }
-        });
-
-
-        return root;
-    }
-
-
-    private static byte[] imageViewToByte(ImageView image) {
-        Bitmap bitmap = ((BitmapDrawable) image.getDrawable()).getBitmap();
-        ByteArrayOutputStream stream = new ByteArrayOutputStream();
-        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream);
-        byte[] byteArray = stream.toByteArray();
-        return byteArray;
-    }
-
-
-    //Pick an image from the gallery
-    private void pickImageFromGallery() {
-        //intent to pick image
-        Intent intent = new Intent(Intent.ACTION_PICK);
-        intent.setType("image/*");
-        startActivityForResult(intent, IMAGE_PICK_CODE);
-    }
-
-    //handle result of runtime permission
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        switch (requestCode) {
-            case PERMISSION_CODE: {
-                if (grantResults.length > 0 && grantResults[0] ==
-                        PackageManager.PERMISSION_GRANTED) {
-                    //permission was granted
-                    pickImageFromGallery();
-                } else {
-                    //permission was denied
-                    Toast.makeText(getActivity(), "Permission denied...!", Toast.LENGTH_SHORT).show();
-                }
-            }
-        }
-    }
-
-    //handle result of picked image
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (resultCode == RESULT_OK && requestCode == IMAGE_PICK_CODE) {
-            //set image to image view
-            mImageView.setImageURI(data.getData());
-            mImageView.setTag(1);
-
-        }
-    }
-}
