+++ app/src/main/res/layout/custom_add_dreamboard.xml
@@ -4,12 +4,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.dream.fragment_addDream">
+    tools:context=".ui.dream.fragment">
 
     <ImageView
         android:id="@+id/image"
         android:layout_width="185dp"
         android:layout_height="185dp"
+        android:layout_marginTop="10dp"
         android:layout_gravity="center"
         android:src="@drawable/ic_outline_photo_library_24"
         android:contentDescription="Image view for adding photo to dreamboard" />
@@ -25,12 +26,13 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="10dp">
+        android:layout_marginTop="30dp">
 
         <TextView
             android:id="@+id/textView2"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
             android:text="Description"
             android:textColor="#151515"
             android:textSize="18sp" />
@@ -38,10 +40,9 @@
             android:id="@+id/description"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginBottom="20dp"
-            android:ems="10"
+            android:layout_marginStart="5dp"
+            android:ems="9"
             android:inputType="textPersonName"
-            android:text="Name" />
+            />
     </LinearLayout>
 </LinearLayout>
