--- app/src/main/res/layout/fragment_add_dream.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".ui.dream.fragment_addDream">
-
-    <ImageView
-        android:id="@+id/image"
-        android:layout_marginTop="100dp"
-        android:layout_width="185dp"
-        android:layout_height="185dp"
-        android:layout_gravity="center"
-        android:src="@drawable/ic_outline_photo_library_24"
-        />
-
-    <TextView
-        android:id="@+id/textView8"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Click to select photo from album"
-        android:textAlignment="center"/>
-
-
-<LinearLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="100dp">
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Description"
-        android:textColor="#151515"
-        android:textSize="18sp" />
-    <EditText
-        android:id="@+id/description"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginBottom="20dp"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="Name" />
-</LinearLayout>
-
-
-    <Button
-        android:id="@+id/add"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:backgroundTint="#4567C5"
-        android:text="Add Photo"
-        android:textColor="#FFFFFF"
-        />
-
-</LinearLayout>
